requirements.txt
schoolopy
pyyaml
matplotlib
numpy